[fix] removed authorisation object
